instructorQuestion is tabbed and has file browser
