+++ lib/question.js
@@ -1072,9 +1072,9 @@ module.exports = {
         if (!assessment) {
             // instructor question pages
             const questionUrl = urlPrefix + '/question/' + question.id + '/';
-            urls.newVariantUrl = questionUrl;
-            urls.tryAgainUrl = questionUrl;
-            urls.reloadUrl = questionUrl + '?variant_id=' + variant.id;
+            urls.newVariantUrl = questionUrl + 'appearance/';
+            urls.tryAgainUrl = questionUrl + 'appearance/';
+            urls.reloadUrl = questionUrl  + 'appearance/' + '?variant_id=' + variant.id;
             urls.clientFilesQuestionUrl = questionUrl + 'clientFilesQuestion';
 
             // necessary for backward compatibility
