+++ pages/instructorQuestionAppearance/instructorQuestionAppearance.ejs
@@ -1,0 +1,61 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <%- include('../partials/head'); %>
+    <script type="text/x-mathjax-config">MathJax.Hub.Config({tex2jax: {inlineMath: [['$','$'], ['\\(','\\)']]}});</script>
+    <script src="/MathJax/MathJax.js?config=TeX-MML-AM_CHTML"></script>
+    <script src="/javascripts/socket.io.js"></script>
+    <script src="/javascripts/d3.min.js"></script>
+    <script src="/javascripts/lodash.min.js"></script>
+    <script src="/localscripts/stacked_histogram.js"></script>
+    <script src="/localscripts/histmini.js"></script>
+    <script>
+      document.urlPrefix = '<%= urlPrefix %>';
+    </script>
+    <% if (question.type != 'Freeform') { %>
+    <script src="/javascripts/lodash.min.js"></script>
+    <script src="/javascripts/require.js"></script>
+    <script src="/localscripts/question.js"></script>
+    <script src="/localscripts/question<%= effectiveQuestionType %>.js"></script>
+    <% } %>
+    <%- extraHeadersHtml %>
+  </head>
+  <body>
+    <script>
+      $(function() {
+        $('[data-toggle="popover"]').popover({
+          sanitize: false
+        });
+        // make the file inputs display the file name
+        $(document).on('change', '.custom-file-input', function () {
+            let fileName = $(this).val().replace(/\\/g, '/').replace(/.*\//, '');
+            $(this).parent('.custom-file').find('.custom-file-label').text(fileName);
+        });
+      });
+    </script>
+    <%- include('../partials/navbar'); %>
+    <div class="container-fluid">
+      <%- include('../partials/navtabs'); %>
+    </div>
+    <div id="content" class="container">
+      <div class="row">
+          <div class="col-lg-9 col-sm-12">
+              <%- include('../partials/question', {question_context: 'instructor'}); %>
+          </div>
+
+          <div class="col-lg-3 col-sm-12">
+              <div class="card mb-4">
+                  <div class="card-header bg-secondary">Student view placeholder</div>
+                  <div class="card-body">
+                      <div class="d-flex justify-content-center">
+                          In student views this area is used for assessment and score info.
+                      </div>
+                  </div>
+              </div>
+              <%- include('../partials/instructorInfoPanel', {question_context: 'instructor'}); %>
+          </div>
+      </div>
+
+    </div>
+  </body>
+</html>
