+++ pages/instructorQuestionOverview/instructorQuestionOverview.sql
@@ -1,30 +1,3 @@
--- BLOCK assessment_question_stats
-SELECT
-    aset.name || ' ' || a.number || ': ' || title AS title,
-    ci.short_name AS course_title,
-    a.id AS assessment_id,
-    a.type AS type,
-    a.course_instance_id,
-    aset.color,
-    (aset.abbreviation || a.number) as label,
-    admin_assessment_question_number(aq.id) as number,
-    aq.*
-FROM
-    assessment_questions AS aq
-    JOIN assessments AS a ON (a.id = aq.assessment_id)
-    JOIN assessment_sets AS aset ON (aset.id = a.assessment_set_id)
-    JOIN course_instances AS ci ON (ci.id = a.course_instance_id)
-WHERE
-    aq.question_id=$question_id
-    AND aq.deleted_at IS NULL
-GROUP BY
-    a.id,
-    aq.id,
-    aset.id,
-    ci.id
-ORDER BY
-    admin_assessment_question_number(aq.id);
-
 -- BLOCK qids
 SELECT
     array_agg(q.qid) AS qids
@@ -79,22 +52,6 @@ FROM
 SELECT
     ci.short_name AS course_instance_directory,
     a.tid AS assessment_directory
-FROM
-    assessment_questions AS aq
-    JOIN assessments AS a ON (a.id = aq.assessment_id)
-    JOIN course_instances AS ci ON (ci.id = a.course_instance_id)
-WHERE
-    aq.question_id = $question_id
-    AND aq.deleted_at IS NULL
-    AND a.deleted_at IS NULL
-    AND ci.deleted_at IS NULL;
-
--- BLOCK select_temp_as_json
-SELECT
-    jsonb_agg(jsonb_build_object(
-        'course_instance_directory', ci.short_name,
-        'assessment_directory', a.tid
-    ))
 FROM
     assessment_questions AS aq
     JOIN assessments AS a ON (a.id = aq.assessment_id)
