+++ pages/partials/instructorInfoPanel.ejs
@@ -4,65 +4,90 @@
   <div class="card-body">
   <h5 class="card-title">User:</h5>
   <div class="d-flex flex-wrap pb-2">
-      <div class="pr-1"><%= user.name %></div>
-      <div class="pr-1"><%= user.uid %></div>
+    <div class="pr-1"><%= user.name %></div>
+    <div class="pr-1"><%= user.uid %></div>
   </div>
 
-      <% if (typeof question !== "undefined") { %>
-          <hr>
-          <h5 class="card-title">Question:</h5>
+  <% if (typeof question !== "undefined") { %>
+  <hr>
+  <h5 class="card-title">Question:</h5>
 
-          <div class="d-flex flex-wrap">
-              <div class="pr-1">QID:</div>
-              <div>
-                  <a href="<%= plainUrlPrefix %>/course_instance/<%= course_instance.id %>/instructor/question/<%= question.id %>?variant_seed=<%= variant.variant_seed %>">
-                      <%= question.qid %></a>
-              </div>
-          </div>
-          <div class="d-flex flex-wrap">
-              <div class="pr-1">Title:</div>
-              <div><%= question.title %></div>
-          </div>
+  <div class="d-flex flex-wrap">
+    <div class="pr-1">QID:</div>
+    <div>
+      <a href="<%= plainUrlPrefix %>/course_instance/<%= course_instance.id %>/instructor/question/<%= question.id %>?variant_seed=<%= variant.variant_seed %>">
+        <%= question.qid %>
+      </a>
+    </div>
+  </div>
+    <div class="d-flex flex-wrap">
+      <div class="pr-1">Title:</div>
+      <div><%= question.title %></div>
+    </div>
+  <div class="d-flex flex-wrap">
+    <div class="pr-1">Started at: </div>
+    <div><%= variant.formatted_date %></div>
+  </div>
+  <div class="d-flex flex-wrap">
+    <div class="pr-1">Duration:</div>
+    <div><%= variant.duration %></div>
+  </div>
+  <div class="d-flex flex-wrap pb-2">
+    <div class="pr-1">
+      <a class="card-link" data-toggle="collapse" href="#instructorTrue_answer">Show/Hide answer</a>
+    </div>
+    <div class="collapse" id="instructorTrue_answer">
+        <code><%= JSON.stringify(variant.true_answer); %></code>
+    </div>
+  </div>
+
+  <% if ((typeof question_context !== "undefined") && (question_context == "instructor")) { %>
+  <div class="row">
+    <div class="col-auto">
+      <button class="btn btn-sm btn-primary" type="button" data-toggle="collapse" data-target="#issueCollapse" aria-expanded="false" aria-controls="issueCollapse">
+        Report an issue with this question
+      </button>
+    </div>
+  </div>
+  <div class="collapse" id="issueCollapse">
+    <hr />
+    <form method="POST">
+      <div class="form-group">
+        <textarea class="form-control" rows="5" name="description" placeholder="Describe the issue"></textarea>
+      </div>
+      <input type="hidden" name="__variant_id" value="<%= variant.id %>">
+      <input type="hidden" name="__csrf_token" value="<%= __csrf_token %>">
+      <div class="form-group text-right">
+        <button class="btn btn-small btn-warning" name="__action" value="report_issue">Report issue</button>
+      </div>
+    </form>
+  </div>
+  <% } %>
+  <% } %>
+
+  <% if (typeof assessment !== "undefined") { %>
+      <hr>
+      <h5 class="card-title">Assessment Instance:</h5>
+      <div class="d-flex flex-wrap">
+          <div class="pr-1">AID:</div>
+          <div><a href="<%= urlPrefix %>/instructor/assessment/<%= assessment.id %>"><%= assessment.tid %></a></div>
+      </div>
+
       <div class="d-flex flex-wrap">
           <div class="pr-1">Started at: </div>
-          <div><%= variant.formatted_date %></div>
+          <div><%= assessment_instance.formatted_date %></div>
       </div>
+
       <div class="d-flex flex-wrap">
           <div class="pr-1">Duration:</div>
-          <div><%= variant.duration %></div>
+          <div><%= assessment_instance.duration %></div>
       </div>
-      <div class="d-flex flex-wrap pb-2">
-          <div class="pr-1">
-              <a class="card-link" data-toggle="collapse" href="#instructorTrue_answer">Show/Hide answer</a>
-          </div>
-          <div class="collapse" id="instructorTrue_answer">
-                <code><%= JSON.stringify(variant.true_answer); %></code>
-          </div>
+
+      <div class="pb-2">
+          <a href="<%= urlPrefix %>/instructor/assessment_instance/<%= assessment_instance.id %>">View log</a>
       </div>
-      <% } %>
+  <% } %>
 
-      <% if (typeof assessment !== "undefined") { %>
-          <hr>
-          <h5 class="card-title">Assessment Instance:</h5>
-          <div class="d-flex flex-wrap">
-              <div class="pr-1">AID:</div>
-              <div><a href="<%= urlPrefix %>/instructor/assessment/<%= assessment.id %>"><%= assessment.tid %></a></div>
-          </div>
-
-          <div class="d-flex flex-wrap">
-              <div class="pr-1">Started at: </div>
-              <div><%= assessment_instance.formatted_date %></div>
-          </div>
-
-          <div class="d-flex flex-wrap">
-              <div class="pr-1">Duration:</div>
-              <div><%= assessment_instance.duration %></div>
-          </div>
-
-          <div class="pb-2">
-              <a href="<%= urlPrefix %>/instructor/assessment_instance/<%= assessment_instance.id %>">View log</a>
-          </div>
-      <% } %>
   </div>
   <div class="card-footer small">
       This box is not visible to students.
