+++ pages/partials/instructorQuestionTabs.ejs
@@ -1,0 +1,14 @@
+<ul class="nav nav-tabs mb-4" id="navtabsInstructorQuestion">
+  <li class="nav-item">
+    <a class="nav-link <% if (navSubPage == "appearance") { %>active<% } %>" href="<%= urlPrefix %>/question/<%= question.id %>/appearance"><i class="fa fa-edit"></i> Appearance</a>
+  </li>
+  <li class="nav-item">
+    <a class="nav-link <% if (navSubPage == "file_view") { %>active<% } %>" href="<%= urlPrefix %>/question/<%= question.id %>/file_view"><i class="fa fa-edit"></i> Files</a>
+  </li>
+  <li class="nav-item">
+    <a class="nav-link <% if (navSubPage == "overview") { %>active<% } %>" href="<%= urlPrefix %>/question/<%= question.id %>/overview"><i class="fa fa-edit"></i> Overview</a>
+  </li>
+  <li class="nav-item">
+    <a class="nav-link <% if (navSubPage == "statistics") { %>active<% } %>" href="<%= urlPrefix %>/question/<%= question.id %>/statistics"><i class="fas fa-chart-bar"></i> Statistics</a>
+  </li>
+</ul>
