+++ pages/partials/navtabs.ejs
@@ -5,5 +5,7 @@
         <%- include('instructorCourseAdminTabs'); %>
     <% } else if (navPage == 'assessment') { %>
         <%- include('instructorAssessmentTabs'); %>
+    <% } else if (navPage == 'question') { %>
+        <%- include('instructorQuestionTabs'); %>
     <% } %>
 <% } %>
